requirements.txt
# apturl==0.5.2
# bcrypt==3.1.4
# beautifulsoup4==4.4.1
# blinker==1.3
# boto==2.49.0
# Brlapi==0.6.4
# certifi==2018.10.15
# cffi==1.11.5
# chardet==3.0.4
# checkbox-support==0.22
# Click==7.0
# command-not-found==0.3
# cryptography==1.2.3
# defer==1.0.6
# defusedxml==0.5.0
Django==2.1.3
django-allauth==0.38.0
django-password-reset==2.0
djangorestframework==3.9.0
Faker==1.0.0
feedparser==5.1.3
guacamole==0.9.2
html5lib==0.999
httplib2==0.9.1
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
language-selector==0.1
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
MarkupSafe==1.1.0
mysqlclient==1.3.14
oauthlib==2.1.0
onboard==1.2.0
padme==1.1.1
# pexpect==4.0.1
# Pillow==3.1.2
# plainbox==0.25
# ptyprocess==0.5
# pyasn1==0.1.9
# pycparser==2.19
# pycups==1.9.73
# pycurl==7.43.0
# pygobject==3.20.0
PyJWT==1.6.4
pyparsing==2.0.3
python-apt==1.1.0b1+ubuntu0.16.4.2
python-dateutil==2.7.5
python-debian==0.1.27
python-social-auth==0.2.1
python-systemd==231
python3-openid==3.1.0
pytz==2018.7
pyxdg==0.25
PyYAML==3.13
#  24538d37a95d716fbe095e7040f18ccd611d7d8a 
# reportlab==3.3.0
# requests==2.20.1
# requests-oauthlib==1.0.0
# sessioninstaller==0.0.0
# six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
SQLAlchemy==1.2.14
# system-service==0.3
# text-unidecode==1.2
# ubuntu-drivers-common==0.0.0
# ufw==0.35
# unattended-upgrades==0.1
# unity-scope-calculator==0.1
# unity-scope-chromiumbookmarks==0.1
# unity-scope-colourlovers==0.1
# unity-scope-devhelp==0.1
# unity-scope-firefoxbookmarks==0.1
# unity-scope-gdrive==0.7
# unity-scope-manpages==0.1
# unity-scope-openclipart==0.1
# unity-scope-texdoc==0.1
# unity-scope-tomboy==0.1
# unity-scope-virtualbox==0.1
# unity-scope-yelp==0.1
# unity-scope-zotero==0.1
# # urllib3==1.24.1
# # usb-creator==0.3.0
# virtualenv==15.0.1
# Werkzeug==0.14.1
# WTForms==2.2.1
# xdiagnose==3.8.4.1
# xkit==0.0.0
# XlsxWriter==0.7.3
